refactor(lint): move print-expression.ts from language/ to rules/gml/
Previously: src/lint/src/language/print-expression.ts
Now: src/lint/src/rules/gml/print-expression.ts

Rationale: The language/ layer owns the ESLint language-adapter and recovery
(gml-language.ts, recovery.ts, path-normalization.ts). print-expression.ts
provides autofix text-generation helpers (printExpression, printNodeForAutofix,
readNodeText) consumed exclusively by two GML lint rules
(optimize-logical-flow-rule.ts, optimize-math-expressions-rule.ts). Neither
gml-language.ts nor recovery.ts imports it, making its placement in language/ a
layer violation. It now lives in rules/gml/ alongside rule-base-helpers.ts.

Updated: language/index.ts (remove re-export), both rule files (import path),
and the test file (import path). Build clean; 565 pass, 3 pre-existing fixture
failures unchanged.
