+++ src/lint/src/language/index.ts
@@ -1,6 +1,5 @@
 export { GML_VISITOR_KEYS, gmlLanguage } from "./gml-language.js";
 export { normalizeLintFilePath } from "./path-normalization.js";
-export { printExpression, printNodeForAutofix, readNodeText } from "./print-expression.js";
 export type {
     InsertedArgumentSeparatorRecovery,
     RecoveryMode,
