+++ src/lint/src/rules/gml/rules/optimize-logical-flow-rule.ts
@@ -1,7 +1,7 @@
 import { Core } from "@gmloop/core";
 import type { Rule } from "eslint";
 
-import { printNodeForAutofix } from "../../../language/index.js";
+import { printNodeForAutofix } from "../print-expression.js";
 import { cloneAstNodeWithoutTraversalLinks, createMeta } from "../rule-base-helpers.js";
 import type { GmlRuleDefinition } from "../rule-definition.js";
 import { applyLogicalNormalizationWithChangeMetadata } from "../transforms/logical-expression-traversal-normalization.js";
